refactor: fix broken component import paths for screens
