+++ App.tsx
@@ -1,13 +1,13 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
-import WelcomeSplash from '@/app/screens/WelcomeSplash';
-import WelcomeWalkThrough from '@/app/screens/WelcomeWalkThrough';
-import OnboardingStepOne from '@/app/screens/OnboardingStepOne';
-import OnboardingStepTwo from '@/app/screens/OnboardingStepTwo';
-import OnboardingStepThree from '@/app/screens/OnboardScreenThree';
-import CreateNewWallet from 'app/CreateNewWallet';
-import ImportExistingWallet from '@/app/screens/ImportExistingWallet';
-import Home from '@/app/screens/Home';
+import WelcomeSplash from '@/app/WelcomeSplash';
+import WelcomeWalkThrough from '@/app/WelcomeWalkThrough';
+import OnboardingStepOne from '@/app/OnboardingStepOne';
+import OnboardingStepTwo from '@/app/OnboardingStepTwo';
+import OnboardingStepThree from '@/app/OnboardScreenThree';
+import CreateNewWallet from '@/app/CreateNewWallet';
+import ImportExistingWallet from '@/app/ImportExistingWallet';
+import Home from '@/app/Home';
 
 
 export default function App() {
